requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.4
django-appconf==1.0.1
django-compressor==2.0
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
drf-nested-routers==0.9.0
gunicorn==19.1.1
Markdown==2.6.5
Pillow==3.1.1
psycopg2==2.6.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
static3==0.5.1
Werkzeug==0.11.4
